+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.13.1...master
+[Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.14.0...master
+
+## [2.14.0] - 2019-08-23
+[2.14.0]: https://github.com/atlassian/aws-infrastructure/compare/release-2.13.1...release-2.14.0
 
 ### Added
 - Let multicast VU nodes share a network. Resolve [JPERF-553].